+++ cardstack/src/components/index.ts
@@ -46,3 +46,4 @@ export * from './MainHeader';
 export * from './TextOptionRow';
 export * from './SwitchSelector';
 export * from './ProgressSteps';
+export * from './ChoosePrepaidCard';
